refactor(users): same schema style as recipes
Refs: #101
