+++ apps/backend/src/modules/users/user.mapper.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 import { createUserDoc } from "@/__tests__/helpers.js";
-import { toUser, toUserSummary } from "./user.mapper.js";
+import { toUserDetails, toUserSummary } from "./user.mapper.js";
 
 describe("toUserSummary", () => {
   it("should map UserSummaryView to UserSummary DTO", () => {
@@ -19,14 +19,14 @@ describe("toUserSummary", () => {
   });
 });
 
-describe("toUser", () => {
+describe("toUserDetails", () => {
   it("should map UserDocument to User DTO", () => {
     const doc = createUserDoc({
       email: "john@test.com",
       name: "John",
     });
 
-    const result = toUser(doc);
+    const result = toUserDetails(doc);
 
     expect(result).toEqual({
       id: doc._id.toString(),
@@ -40,7 +40,7 @@ describe("toUser", () => {
   it("should not expose password field", () => {
     const doc = createUserDoc();
 
-    const result = toUser(doc);
+    const result = toUserDetails(doc);
 
     expect(result).not.toHaveProperty(["_id", "__v", "password", "role"]);
   });
