+++ packages/shared/src/recipes/recipe.response.schema.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import { categorySummarySchema } from "../categories/category.schema.js";
 import { persistenceFieldsSchema } from "../common/persistence.schema.js";
-import { userSummarySchema } from "../users/user.schema.js";
+import { userSummarySchema } from "../users/user.response.schema.js";
 import { createRecipeInputSchema } from "./recipe.input.schema.js";
 import { recipeStatsSchema } from "./recipe.schema.js";
