+++ packages/shared/src/reviews/review.response.schema.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { persistenceFieldsSchema } from "../common/persistence.schema.js";
-import { userSummarySchema } from "../users/user.schema.js";
+import { userSummarySchema } from "../users/user.response.schema.js";
 import { createReviewInputSchema } from "./review.input.schema.js";
 
 export const reviewsStatsSchema = z.object({
